refactor: move safeguard to feature env level
