+++ frontend/src/component/feature/FeatureView/FeatureOverview/ReleasePlan/ChangeRequest/ReleasePlanChangeRequestDialog.tsx
@@ -27,10 +27,6 @@ type ChangeRequestAction =
     | {
           type: 'deleteMilestoneProgression';
           sourceMilestoneId: string;
-      }
-    | {
-          type: 'deleteSafeguard';
-          safeguardId: string;
       };
 
 interface IReleasePlanChangeRequestDialogProps {
@@ -129,17 +125,6 @@ export const ReleasePlanChangeRequestDialog = ({
                     </p>
                 );
             }
-
-            case 'deleteSafeguard':
-                return (
-                    <p>
-                        <StyledBoldSpan>Remove</StyledBoldSpan> safeguard from
-                        release plan{' '}
-                        <StyledBoldSpan>{releasePlan.name}</StyledBoldSpan> for{' '}
-                        <StyledBoldSpan>{featureId}</StyledBoldSpan> in{' '}
-                        <StyledBoldSpan>{environmentId}</StyledBoldSpan>
-                    </p>
-                );
         }
     };
